Fix references to renamed folders
