+++ packages/wallet/src/redux/__tests__/helpers.ts
@@ -1,4 +1,4 @@
-import { ChannelStatus } from '../channelState/state';
+import { ChannelStatus } from '../channel-state/state';
 import { StateWithSideEffects } from '../utils';
 import { Commitment } from 'fmg-core';
 import { TransactionOutboxItem } from '../outbox/state';
