+++ packages/wallet/src/redux/indirect-funding/reducer.ts
@@ -3,7 +3,7 @@ import * as indirectFundingState from './state';
 import * as actions from '../actions';
 import { unreachable } from '../../utils/reducer-utils';
 import { PlayerIndex } from '../types';
-import { playerAReducer } from './playerA/reducer';
+import { playerAReducer } from './player-a/reducer';
 
 export const indirectFundingReducer = (
   state: states.Initialized,
